PR for update to v.1.0.5

+++ controllers/cards.js
@@ -1,25 +1,39 @@
 const path = require('path');
 // eslint-disable-next-line import/no-dynamic-require
 const Card = require(path.join('..', 'models', 'card'));
-const NotFoundError = new Error();
+const customError = new Error();
 
 module.exports.deleteCard = (req, res) => {
-  Card.findByIdAndRemove(req.params._id)
+  Card.findById(req.params)
     .then((card) => {
       if (!card) {
-        NotFoundError.message = 'Карточки с указанным id не существует или была удалена';
-        NotFoundError.name = 'NotFoundError';
-        throw NotFoundError;
+        customError.message = 'Карточки с указанным id не существует';
+        customError.name = 'NotFoundError';
+        throw customError;
       }
-      res.status(200).send({ message: `Карточка ${card._id} успешно удалена` });
+      if (card.owner.toString() !== req.user._id) {
+        customError.message = 'Недостаточно прав';
+        customError.name = 'AuthError';
+        throw customError;
+      }
+      return Card.findByIdAndRemove(req.params._id)
+        .then(() => {
+          res.status(200).send({ message: 'Карточка успешно удалена' });
+        });
     })
     .catch((err) => {
-      if (err.name === 'CastError') {
-        res.status(400).send({ message: err.message });
-      } else if (err.name === 'NotFoundError') {
-        res.status(404).send({ message: err.message });
-      } else {
-        res.status(500).send({ message: err.message });
+      switch (err.name) {
+        case 'CastError':
+          res.status(400).send({ message: `${err.name}: Ошибка запроса` });
+          break;
+        case 'NotFoundError':
+          res.status(404).send({ message: err.message });
+          break;
+        case 'AuthError':
+          res.status(403).send({ message: err.message });
+          break;
+        default:
+          res.status(500).send({ message: err.message });
       }
     });
 };
@@ -64,15 +78,15 @@ module.exports.likeCard = (req, res) => {
   )
     .then((card) => {
       if (!card) {
-        NotFoundError.message = 'Карточка не найдена';
-        NotFoundError.name = 'NotFoundError';
-        throw NotFoundError;
+        customError.message = 'Карточка не найдена';
+        customError.name = 'NotFoundError';
+        throw customError;
       }
       res.status(200).send({ _id: req.params.cardId, likes: card.likes.length });
     })
     .catch((err) => {
       if (err.name === 'CastError') {
-        res.status(400).send({ message: err.message });
+        res.status(400).send({ message: `${err.name}: Ошибка запроса` });
       } else if (err.name === 'NotFoundError') {
         res.status(404).send({ message: err.message });
       } else {
@@ -89,15 +103,15 @@ module.exports.dislikeCard = (req, res) => {
   )
     .then((card) => {
       if (!card) {
-        NotFoundError.message = 'Карточка не найдена';
-        NotFoundError.name = 'NotFoundError';
-        throw NotFoundError;
+        customError.message = 'Карточка не найдена';
+        customError.name = 'NotFoundError';
+        throw customError;
       }
       res.status(200).send({ _id: req.params.cardId, likes: card.likes.length });
     })
     .catch((err) => {
       if (err.name === 'CastError') {
-        res.status(400).send({ message: err.message });
+        res.status(400).send({ message: `${err.name}: Ошибка запроса` });
       } else if (err.name === 'NotFoundError') {
         res.status(404).send({ message: err.message });
       } else {
